Constructor of S3 objects of class master_selection

Constructor of S3 objects of class master_selection

new_master_selection(data_matrix, preselected_sites = NULL, region,
                     mask = NULL, raster_base, PCA_results = NULL,
                     selected_sites_random = NULL, selected_sites_G = NULL,
                     selected_sites_E = NULL, selected_sites_EG = NULL)

Arguments

data_matrix
date.frame with information about geographic location of raster cells, initial environmental data, and if available, the first two principal components derived from an analysis done with environmental values.

preselected_sites
data.frame containing sites that must be included in posterior selections of sites for the survey system. Columns must be: "Sites", "Longitude", "Latitude", in that order. Default = NULL.

region
SpatialPolygons* object representing the region of interest.

mask
SpatialPolygons* object used. Default = NULL.

raster_base
a SpatialPolygonsDataFrame representing the grid of the raster layers used, which will be used for plotting purposes.

PCA_results
results of principal component analysis performed with values from raster layers used. Default = NULL.

selected_sites_random
data.frame with the sites selected randomly. Default = NULL.

selected_sites_G
data.frame with the sites selected based on geographic distances. Default = NULL.

selected_sites_E
data.frame with the sites selected based on environmental distances. Default = NULL.

selected_sites_EG
data.frame with the sites selected based on environmental and geographic considerations. Default = NULL.

Value

An S3 object of class master_selection.