Dissimilarity indices from PAM_subset

Computes dissimilarity indices for each set of selected sites contained in elements of PAM_subset that contain information of species incidence (presence-absence). Calculations are done also among sets of selected sites.

selected_sites_DI(PAM_subset, selection_type = "all", method = "jaccard",
                  verbose = TRUE, ...)

Arguments

PAM_subset
object of class PAM_subset obtained using the function subset_PAM.

selection_type
type of selection to be considered when creating dissimilarity matrices for elements in PAM_subset. Options are: "all", "random", "E", "G", and "EG". The default, "all", uses all selection types present in PAM_subset.

method
(character) dissimilarity index to be passed to function vegdist. Default = "jaccard". See details.

verbose
(logical) whether or not to print messages about the process. Default = TRUE.

...
other arguments to be passed to function vegdist.

Value

A list containing:

Dissimilarity matrices for all PAMs reduced based on distinct sets of selected sites.

A matrix summarizing incidences from all sets of selected sites.

A dissimilarity matrix for the summary of incidences for all sets of selected sites.

The result of clustering sets of selected sites based on dissimilarities.

Details

Important details about the process performed to compute dissimilarity indices can be seen in the documentation of vegdist.

Examples

# Data
data("b_pam", package = "biosurvey")
data("m_selection", package = "biosurvey")

# Subsetting base PAM according to selections
sub_pam_all <- subset_PAM(b_pam, m_selection, selection_type = "all")

# Calculating dissimilarities
DI_sel <- selected_sites_DI(sub_pam_all)

#> Running analysis...

#> Random selection

#> 	One or more sites were excluded due to lack of species data:
#> 	Site_12

#> G selection

#> 	One or more sites were excluded due to lack of species data:
#> 	Site_1, Site_8, Site_10, Site_15

#> E selection

#> Summary of all selections