Species accumulation curves from PAM_subset

Creates species accumulation curves for each set of selected sites contained in elements of PAM_subset that contain information of species incidence (presence-absence).

selected_sites_SAC(PAM_subset, selection_type = "all", method = "exact", ...)

Arguments

PAM_subset
object of class PAM_subset obtained using the function subset_PAM.

selection_type
type of selection to be considered when creating SAC for elements in PAM_subset. Options are: "all", "random", "E", "G", #' and "EG". The default, "all", uses all selection types present in PAM_subset.

method
(character) species accumulation method to be passed to function specaccum. Default = "exact".

...
other arguments to be passed to function specaccum.

Value

A list of species accumulation curves (SACs, "specaccum" objects) for all sets of selected sites according to option defined in selection_type.

Details

Important details about the process performed to obtain each of the SACs can be seen in the help for function specaccum.

Examples

# Data
data("b_pam", package = "biosurvey")
data("m_selection", package = "biosurvey")

# Subsetting base PAM according to selections
sub_pam_all <- subset_PAM(b_pam, m_selection, selection_type = "all")

SACs <- selected_sites_SAC(PAM_subset = sub_pam_all, selection_type = "all")