Example of stack of layers of suitable and unsuitable conditions for species

An object of class RasterStack containing information about suitable and unsuitable conditions for five species.

Format

RasterStack with 109 rows, 182 columns, 19838 cells, and 5 layers:

RasterLayer

suitable (1) and unsuitable (0) conditions

Examples

sp_layers <- raster::stack(system.file("extdata/sp_layers.tif",
                           package = "biosurvey"))

sp_layers

#> class      : RasterStack 
#> dimensions : 109, 182, 19838, 5  (nrow, ncol, ncell, nlayers)
#> resolution : 0.1666667, 0.1666667  (x, y)
#> extent     : -117.1324, -86.79904, 14.56388, 32.73055  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> names      : sp_layers.1, sp_layers.2, sp_layers.3, sp_layers.4, sp_layers.5 
#> min values :           0,           0,           0,           0,           0 
#> max values :           1,           1,           1,           1,           1 
#>