Biodiversity indices derived from PAM

Calculates a set of biodiversity indices using values contained in a presence-absence matrix.

PAM_indices(PAM, indices = "all", exclude_column = NULL)

Arguments

PAM
matrix, data.frame, or base_PAM object containing information on species presence and absence for a set of sites. Sites are organized in the rows and species in the columns. See details.

indices
(character) code for indices to be calculated. Basic indices are calculated all the time; other indices need to be specified. Options are: "all", "basic, "AB", "BW", "BL", "SCSC", "SCSR", "DF", "CC", "WRN", "SRC", "CMSC", "CMSR", "MCC", and "MRC". Default = "all". See details.

exclude_column
(optional) name or numeric index of columns to be excluded. Default = NULL.

Value

If PAM is a matrix or data.frame, the result is a list with the results described below (depending on indices). If PAM is a base_PAM object, a base_PAM object will be returned and the list described above will be appended to the element PAM_indices in such an element.

Details

Descriptions of the codes of all indices to be calculated are presented in the table below. If indices = "basic", only basic indices are calculated. However, basic indices are calculated in all cases not matter the code(s) defined in indices. Some indices require previous calculations of other indices, in such cases, all indices required are added to the final list. For further details on the way calculations are performed and the meaning of the indices see Soberón and Cavner (2015) doi: 10.17161/bi.v10i0.4801 .

Code	Index	Calculation

RI	Richness	Basic

RA	Range	Basic

RIN	Richness normalized	Basic

RAN	Range normalized	Basic

AB	Additive Beta	Needs to be defined

BW	Beta Whittaker	Needs to be defined

BL	Beta Legendre	Needs to be defined and DF

SCSC	Schluter covariance sites-composition	Needs to be defined and CMSC

SCSR	Schluter covariance species-ranges	Needs to be defined and CMSR

DF	Dispersion field	Needs to be defined

SCC	Shared community composition	Needs to be defined

WRN	Wright-Reeves nestedness	Needs to be defined, BW, and DF

SRC	Stone-Roberts C-score	Needs to be defined and DF

CMSC	Covariance matrix sites-composition	Needs to be defined, DF, and BW

CMSR	Covariance matrix species-ranges	Needs to be defined, SCC, and BW

MCC	Mean composition covariance	Calculated with CMSC

MRC	Mean range covariance	Calculated with CMSR

See also

Examples

# Data
data("sp_data", package = "biosurvey")

# PAM
pam <- PAM_from_table(data = sp_data, ID_column = "ID",
                      species_column = "Species")

pam_ind <- PAM_indices(pam, exclude_column = 1)
pam_ind[1:3]

#> $One_value_indices
#>                                         Value
#> Sites_Cells                        173.000000
#> Species                             25.000000
#> Av_dispersion_field                 90.450867
#> Av_shared_community_composition    101.120000
#> Additive_Beta                       21.589595
#> Beta_Whittaker                       7.330508
#> Beta_Legendre                   -15058.000000
#> Schluter_cov_sites_composition      36.741123
#> Schluter_cov_species_ranges          3.971186
#> Wright_Reeves_nestedness          7529.000000
#> Stone_Roberts_Cscore             25140.000000
#> 
#> $Richness
#>   5  35  36  38  39  40  41  42  43  44  45  68  71  72  73  74  75  76  77  78 
#>   1   1   1   1   1   1   1   2   2   2   1   1   1   2   2   2   2   2   6   5 
#> 100 101 103 104 105 106 107 108 109 110 111 112 113 133 134 136 137 138 139 140 
#>   1   1   2   2   2   2   2   2   6   5   6   6   5   1   1   2   1   2   2   2 
#> 141 142 143 144 145 146 166 169 170 171 172 173 174 175 176 177 178 179 202 203 
#>   4   6   6   6   6   5   1   1   2   2   2   4   4   6   6   5   5   4   1   2 
#> 204 205 206 207 208 209 210 211 212 235 236 237 238 239 240 241 242 243 244 245 
#>   2   3   3   2   3   2   2   4   4   3   3   3   3   3   3   2   3   6   6   6 
#> 264 268 269 270 271 272 273 274 275 276 277 297 301 302 303 304 305 306 307 308 
#>   1   3   3   3   2   4   3   4   6   6   6   1   3   2   3   4   3   7   6   6 
#> 309 334 335 336 337 338 339 340 341 366 367 368 369 370 371 372 373 382 383 398 
#>   6   2   3   4   3   4   6   7   5   1   2   4   3   3   5   6   5   4   2   1 
#> 399 400 401 402 403 404 405 406 413 414 415 431 432 433 434 435 436 437 438 444 
#>   2   3   4   4   4   5   6   5   3   2   1   1   2   3   4   4   6   6   6   5 
#> 445 446 447 464 465 466 467 468 469 470 471 472 473 474 475 476 477 499 500 501 
#>   3   2   2   2   2   3   4   5   5   6   6   5   5   5   5   4   3   2   4   4 
#> 502 503 504 505 506 507 533 534 535 537 538 539 540 
#>   5   6   6   5   5   5   1   3   4   3   2   3   3 
#> 
#> $Range
#>  Species_1  Species_2  Species_3  Species_4  Species_5  Species_6  Species_7 
#>         77         19          5         20         16         18         17 
#>  Species_8  Species_9 Species_10 Species_11 Species_12 Species_13 Species_14 
#>         14         22         16         21         20         14         16 
#> Species_15 Species_16 Species_17 Species_18 Species_19 Species_20 Species_21 
#>         18         12         18         15         15         19         14 
#> Species_22 Species_23 Species_24 Species_25 
#>        140         12         19         13 
#>