biosurvey: Tools for Biological Survey Planning

biosurvey is a collection of tools that allow users to plan sampling sites. The methods presented increase the efficiency of biodiversity monitoring by considering the relationship between environmental and geographic conditions in a region.

Details

Three main modules are included: 1) data preparation, 2) selection of sets of sites for biodiversity sampling, and 3) tools for testing efficiency of distinct sets of sampling sites. Data are prepared in ways that avoid the need for more data in posterior analyses, and allow concentrating in critical methodological decisions to select sampling sites. Various algorithms for selecting sampling sites are available, and options for considering pre-selected sites (known to be important for biodiversity monitoring) are included. Visualization is a critical component in this set of tools and most of the results obtained can be plotted to help to understand their implications. The options for selecting sampling sites included here differ from other implementations in that they consider the environmental and geographic structure of a region to suggest sampling sites that could increase the efficiency of efforts dedicated to monitoring biodiversity.

Main functions in biosurvey

block_sample, compare_SAC, EG_selection, DI_dendrogram, explore_data_EG, make_blocks, PAM_indices, plot_blocks_EG, plot_DI, plot_PAM_geo, plot_PAM_CS, plot_SAC, plot_sites_EG, prepare_base_PAM, prepare_master_matrix, prepare_PAM_CS random_selection, selected_sites_DI selected_sites_SAC, subset_PAM, uniformE_selection, uniformG_selection

Other functions (important helpers)

assign_blocks, closest_to_centroid, distance_filter, files_2data, find_clusters, find_modes, grid_from_region, match_rformat, PAM_from_table, point_sample, point_sample_cluster, point_thinning, refill_PAM_indices, rlist_2data, selected_sites_PAM, spdf_2data, stack_2data, unimodal_test, wgs84_2aed_laea

Data included

b_pam, distance_filter, dist_list, files_2data, m_matrix, m_matrix_pre, m_selection, mx, preselected, sp_data, species_data, sp_layers, sp_occurrences, variables,