Creates a data.frame of species' references from files in a directory

Creates a data.frame of species' references that contains longitude, latitude, and species name, from a character.

files_2data(path, format, spdf_grid = NULL, parallel = FALSE, n_cores = NULL)

Arguments

path
(character) full path name of directory containing raster, shapefiles, geopackage, or GeoJSON files representing species geographic ranges. Each file must be named as the species that it represents. All files must be in the same format. If files are raster, values in each layer must be 1 (presence) and 0 (absence).

format
(character) the format files found in path. Current available formats are: "shp", "gpkg", "geojson", "GTiff", and "ascii".

spdf_grid
geographic grid for the region of interest (output of function grid_from_region). Used when format equals "shp" or "gpkg". Default = NULL.

parallel
(logical) whether to perform analyses in parallel. Default = FALSE.

n_cores
(numeric) number of cores to be used when parallel = TRUE. The default, NULL, uses available cores - 1.

Value

If files are in raster format, a data.frame of species geographic records derived from values of presence in each layer.

If files are not in raster format, a data.frame of species' found in distinct positions (defined with identifiers); includes two columns: "ID" and "Species".

Examples

# \donttest{
# Data for examples
data("mx", package = "biosurvey")
data("species_data", package = "biosurvey")

# Saving species data in a temporal directory
tdir <- file.path(tempdir(), "testbio")
dir.create(tdir)

namessp <- paste0("species_", 1:length(species_data))


for (i in 1:length(species_data)) {
  rgdal::writeOGR(species_data[i, ], dsn = tdir, layer = namessp[i],
                  driver = "ESRI Shapefile")
}

# Preparing grid for analysis
grid_reg <- grid_from_region(region = mx, cell_size = 100)

#> Warning: CRS object has comment, which is lost in output

#> Warning: CRS object has comment, which is lost in output


# Running analysis with data from directory
sp_data <- files_2data(path = tdir, format = "shp", spdf_grid = grid_reg)

#>   |                                                                              |                                                                      |   0%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |===                                                                   |   4%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |======                                                                |   8%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=========                                                             |  12%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |============                                                          |  17%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |===============                                                       |  21%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |==================                                                    |  25%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |====================                                                  |  29%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=======================                                               |  33%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |==========================                                            |  38%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=============================                                         |  42%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |================================                                      |  46%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |===================================                                   |  50%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |======================================                                |  54%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=========================================                             |  58%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |============================================                          |  62%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |===============================================                       |  67%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |==================================================                    |  71%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |====================================================                  |  75%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=======================================================               |  79%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |==========================================================            |  83%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |=============================================================         |  88%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |================================================================      |  92%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |===================================================================   |  96%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#>   |                                                                              |======================================================================| 100%

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> 

# }