Detection of clusters in 2D spaces

Finds clusters of data in two dimensions based on distinct methods.

find_clusters(data, x_column, y_column, space,
              cluster_method = "hierarchical", n_k_means = NULL,
              split_distance = NULL)

Arguments

data
matrix or data.frame that contains at least two columns.

x_column
(character) the name of the x-axis.

y_column
(character) the name of the y-axis.

space
(character) space in which the thinning will be performed. There are two options available: "G", if it will be in the geographic space, and "E", if it will be in the environmental space.

cluster_method
(character) name of the method to be used for detecting clusters. Options are "hierarchical" and "k-means"; default = "hierarchical".

n_k_means
(numeric) number of clusters to be identified when using the "k-means" in cluster_method.

split_distance
(numeric) distance in meters (if space = "G") or Euclidean distance (if space = "E") to identify clusters if cluster_method = "hierarchical".

Value

A data frame containing data and an additional column defining clusters.

Details

Clustering methods make distinct assumptions and one of them may perform better than the other depending on the pattern of the data.

The k-means method tends to perform better when data are grouped spatially (spherically) and clusters are of a similar size. The hierarchical clustering algorithm usually takes more time than the k-means method. Both methods make assumptions and may work well on some data sets but fail on others.

Examples

# Data
data("m_matrix", package = "biosurvey")

# Cluster detection
clusters <-  find_clusters(m_matrix$data_matrix, x_column = "PC1",
                           y_column = "PC2", space = "E",
                           cluster_method = "hierarchical", n_k_means = NULL,
                           split_distance = 4)
head(clusters)

#>   Longitude Latitude Mean_temperature Max_temperature Min_temperature
#> 1 -116.4167 32.58333              146             329               9
#> 2 -116.2500 32.58333              148             333               7
#> 3 -116.0833 32.58333              155             342               9
#> 4 -115.9167 32.58333              198             387              37
#> 5 -115.7500 32.58333              215             403              46
#> 6 -115.5833 32.58333              221             412              45
#>   Annual_precipitation Prec_wettest_month Prec_driest_month       PC1
#> 1                  400                 76                 1 -2.052589
#> 2                  320                 60                 1 -2.170276
#> 3                  237                 39                 1 -2.213166
#> 4                  235                 45                 0 -1.421165
#> 5                  233                 45                 0 -1.129137
#> 6                  148                 25                 0 -1.203997
#>          PC2 clusters
#> 1 -0.4190674        1
#> 2 -0.2378053        1
#> 3  0.1379122        1
#> 4  1.6995256        2
#> 5  2.2733643        2
#> 6  2.6227359        2